[HW-1]: Update Task1
* replace scanner.hasNextFloat() with scanner.hasNextDouble() in main() method;
* replace scanner.nextFloat() with scanner.nextDouble() in main() method.

+++ java/Oleksii-Riabov/src/com/zoolatech/lecture1/tasks/_1/Task1.java
@@ -12,8 +12,8 @@ public class Task1 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        if(scanner.hasNextFloat()){
-            milesInKilometers(scanner.nextFloat());
+        if(scanner.hasNextDouble()){
+            milesInKilometers(scanner.nextDouble());
         } else {
             throw new IllegalArgumentException("Enter numeric value");
         }
